+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- `sicd_area_plane` module
 - `sicd_chip`
 - `sicd_deskew` module
 - `sicd_scene_to_image`